new useSearchTerms hook for getting search terms from redux search form, deleting a single search term or deleting all search terms, moved buildSearchParams into /utils directory
